refactor: Rename Compose Core module into Compose Basics
