+++ Core2/Compose/Basics/proguard-rules.pro
