+++ Core2/Compose/Basics/src/main/java/com/infomaniak/core2/compose/basics/ScreenUtils.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.infomaniak.core2.compose.core
+package com.infomaniak.core2.compose.basics
 
 import android.annotation.SuppressLint
 import android.app.Activity
