+++ app/src/main/java/com/infomaniak/swisstransfer/ui/OnboardingActivity.kt
@@ -27,7 +27,7 @@ import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.rememberCoroutineScope
-import com.infomaniak.core2.compose.core.LockScreenOrientation
+import com.infomaniak.core2.compose.basics.LockScreenOrientation
 import com.infomaniak.swisstransfer.ui.screen.onboarding.OnboardingScreen
 import com.infomaniak.swisstransfer.ui.theme.LocalWindowAdaptiveInfo
 import com.infomaniak.swisstransfer.ui.theme.SwissTransferTheme
